# src/main/resources/swagger/menuguru.yaml
openapi: 3.0.3
info:
  title: MenuGuru - Application API
  description: Fictitious order management system (MenuGuru) for restaurants as part of the evaluation work for the Postgraduate course in Software Architecture at FIAP. MenuGuru's main objective is to record and track the status of orders for restaurants, where the customer can place their order and track its status until pickup.
  version: 1.0.0
...
servers:
  - url: /menuguru/v1
...
paths:
...
  /products:
    post:
      tags:
        - products
      summary: Create new product
      description: Create a new product
      operationId: createProduct
      requestBody:
        description: Create a new product
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/Product'
              required:
                - name
                - productType
                - description
                - price
                - image
        required: true        
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
          headers: 
            Location:
              description: Location of the new product register
              schema:
                type: string
                example: /products/1
        '400':
            description: Bad request
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorTemplate'
                example:
                  type: about:blank
                  tittle: Bad request
                  detail: Malformed request or unknown fields
                  instance: /products
        '406':
            description: Not acceptable
        '415':
            description: Unsupported Media Type
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorTemplate'
                example:
                  type: about:blank
                  tittle: Unsupported Media Type
                  detail: Content-Type 'text/plain' is not supported
                  instance: /products
        '422':
            description: Unprocessable entity
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorTemplate'
                example:
                  type: about:blank
                  tittle: Unprocessable entity
                  detail: Invalid product type
                  instance: /products
        '500':
            description: Internal Server Error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorTemplate'
                example:
                  type: about:blank
                  tittle: Internal Server Error
                  detail: A error ocurred during the operation. Please refer to application log for details
                  instance: /products
    get:
      tags:
        - products
      summary: List products
      description: List all products using pagination criteria
      operationId: listProducts
      parameters: 
        - name: _page
          description: Defines the page number that will be retrieved
          in: query
          required: true
          example: 1
          schema:
              $ref: '#/components/schemas/PageParameter'
        - name: _size
          description: Defines the page size (rows number) that will be retrieved
          in: query
          required: true
          example: 50
          schema:
              $ref: '#/components/schemas/SizeParameter'
        - name: type
          description: Categorize the product according with his type.
          in: query
          required: false
          schema:
              $ref: '#/components/schemas/ProductType'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageTemplate'
                  - type: object
                    properties:
                      _content:
                        description: Itens retrieved. If there is no item the list will be empty
                        type: array
                        items:
                          allOf:
                          - $ref: '#/components/schemas/Product'
        '400':
            description: Bad request
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorTemplate'
                example:
                  type: about:blank
                  tittle: Bad request
                  detail: Malformed request or unknown fields
                  instance: /products
        '406':
            description: Not acceptable
        '415':
            description: Unsupported Media Type
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorTemplate'
                example:
                  type: about:blank
                  tittle: Unsupported Media Type
                  detail: Content-Type 'text/plain' is not supported
                  instance: /products
        '422':
            description: Unprocessable entity
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorTemplate'
                example:
                  type: about:blank
                  tittle: Unprocessable entity
                  detail: 'Invalid size parameter. Value must be greater than 0 and less than 50. Actual value: 51'
                  instance: /products
        '500':
            description: Internal Server Error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorTemplate'
                example:
                  type: about:blank
                  tittle: Internal Server Error
                  detail: A error ocurred during the operation. Please refer to application log for details
                  instance: /products      
  /products/{productId}:
    get:
      tags:
        - products
      summary: Get product
      description: Retrieve a specific product according with id
      operationId: getProduct
      parameters: 
        - name: productId
          description: Unique key that identifies the product
          in: path
          required: true
          example: 1
          schema:
              type: integer
              minimum: 1
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
            description: Bad request
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorTemplate'
                example:
                  type: about:blank
                  tittle: Bad request
                  detail: Malformed request or unknown fields
                  instance: /products
        '404':
            description: Resource not found
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorTemplate'
                example:
                  type: about:blank
                  tittle: Resource not found
                  detail: Product not found
                  instance: /products/2
        '406':
            description: Not acceptable
        '415':
            description: Unsupported Media Type
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorTemplate'
                example:
                  type: about:blank
                  tittle: Unsupported Media Type
                  detail: Content-Type 'text/plain' is not supported
                  instance: /products
        '500':
            description: Internal Server Error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorTemplate'
                example:
                  type: about:blank
                  tittle: Internal Server Error
                  detail: A error ocurred during the operation. Please refer to application log for details
                  instance: /products
    patch:
      tags:
        - products
      description: Updates the properties of the specific
      operationId: updateProduct
      parameters: 
        - name: productId
          description: Unique key that identifies the product
          in: path
          required: true
          example: 1
          schema:
              type: integer
              minimum: 1      
      requestBody:
        description: Updates a product
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
            description: Bad request
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorTemplate'
                example:
                  type: about:blank
                  tittle: Bad request
                  detail: Malformed request or unknown fields
                  instance: /products
        '404':
            description: Resource not found
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorTemplate'
                example:
                  type: about:blank
                  tittle: Resource not found
                  detail: Product not found
                  instance: /products/2              
        '406':
            description: Not acceptable
        '415':
            description: Unsupported Media Type
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorTemplate'
                example:
                  type: about:blank
                  tittle: Unsupported Media Type
                  detail: Content-Type 'text/plain' is not supported
                  instance: /products
        '500':
            description: Internal Server Error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorTemplate'
                example:
                  type: about:blank
                  tittle: Internal Server Error
                  detail: A error ocurred during the operation. Please refer to application log for details
                  instance: /products                  
    delete:
      tags:
        - products
      summary: Delete customer
      description: Delete a specific product according with id. The operation can't be undone
      operationId: deleteProduct
      parameters: 
        - name: productId
          description: Unique key that identifies the customer
          in: path
          required: true
          example: 1
          schema:
              type: integer
              minimum: 1
      responses:
        '204':
          description: No content
        '404':
            description: Resource not found
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorTemplate'
                example:
                  type: about:blank
                  tittle: Resource not found
                  detail: Customer not found
                  instance: /products/2
        '406':
            description: Not acceptable
        '415':
            description: Unsupported Media Type
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorTemplate'
                example:
                  type: about:blank
                  tittle: Unsupported Media Type
                  detail: Content-Type 'text/plain' is not supported
                  instance: /products
        '500':
            description: Internal Server Error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorTemplate'
                example:
                  type: about:blank
                  tittle: Internal Server Error
                  detail: A error ocurred during the operation. Please refer to application log for details
                  instance: /products
...
components:
  schemas:
...
    ProductType:
      description: Categorize the product according with his type.
      type: string
      enum:
        - LANCHE
        - ACOMPANHAMENTO
        - BEBIDA
        - SOBREMESA
    TypeIdProduct:
      description: Unique key that identifies the product
      type: integer
      format: int64
      example: 10
    TypeProductPrice:
      description: Value of the product that should be payed to achieve him.
      type: number
      format: double
      minimum: 0.01
      example: 39.90    
    Product:
      type: object
      description: A food item selled in the restaurant
      properties:
        id:
          allOf:
            - $ref: "#/components/schemas/TypeIdProduct"
          readOnly: true
        name:
          description: Word or expression that designates the product.
          type: string
          minLength: 1
          maxLength: 100
          example: X-Bacon
        productType:
          $ref: "#/components/schemas/ProductType"
        description:
          description: Short text that describes the product and his qualities
          type: string
          minLength: 1
          maxLength: 250
          example: O x-bacon é um sanduíche irresistível que une o sabor intenso do bacon crocante com queijo derretido, alface, tomate e um suculento hambúrguer, tudo envolto em um pão macio e tostado. Uma explosão de sabores em cada mordida!
        price:
          $ref: "#/components/schemas/TypeProductPrice"
        image:
          description: Visual representation of the product
          type: string
          format: base64
          example: iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mNk+A8AAQUBAScY42YAAAAASUVORK5CYII=
...
    PageParameter:
      type: integer
      minimum: 1
    SizeParameter:
      type: integer
      minimum: 1
      maximum: 50
    PageTemplate:
      type: object
      properties:
          _page:
              type: integer
              description: Actual page retrieved
              example: 1
          _size:
              type: integer
              description: Page size (rows number) requested
              example: 50
          _pageElements:
              type: integer
              description: Actual page size (rows number) retrieved
              example: 1
    ErrorTemplate:
      description: Problem detail objects acording to [RFC9457](https://www.rfc-editor.org/rfc/rfc9457)
      type: object
      properties:
        type:
          type: string
          description: Containing a URI reference that identifies the problem type. When this member is not present, its value is assumed to be "about:blank".
          example: https://example.net/validation-error
        tittle:
          type: string
          description: Contain a short, human-readable summary of the problem type.
          example: Your request is not valid
        detail:
            type: string
            description: A human-readable explanation specific to this occurrence of the problem.
            example: Customer name must be not greater than 100 characters.
        instance: 
            type: string
            description: String containing a URI reference that identifies the specific occurrence of the problem
            example: /customers/1